Finish C library name transition.
Remove all internal macros and cast various invocations as
necessary. Jung-uk Kim, Bob Moore, Lv Zheng.

+++ source/components/tables/tbprint.c
@@ -156,7 +156,7 @@ AcpiTbFixString (
 
     while (Length && *String)
     {
-        if (!isprint (*String))
+        if (!isprint ((int) *String))
         {
             *String = '?';
         }
